intelligence-service/requirements.txt
fastapi
uvicorn
scikit-learn
pandas
joblib
psycopg2-binary
python-dotenv
# OCR Dependencies
pydantic-settings
python-multipart
pdfplumber
loguru
